split api and admin update validators and improve test coverage
